move integration test function to integration test
