+++ core/tests/agnostic_tests.rs
@@ -1,7 +1,36 @@
+use std::io::Write;
 use typeshare_core::{
-    language::TypeScript, parser::ParseError, process_input, rust_types::RustTypeParseError,
+    language::{CrateTypes, Language, TypeScript},
+    parser::{self, ParseError},
+    rust_types::RustTypeParseError,
     ProcessInputError,
 };
+/// Parse and generate types for a single Rust input file.
+pub fn process_input(
+    input: &str,
+    language: &mut dyn Language,
+    imports: &CrateTypes,
+    out: &mut dyn Write,
+) -> Result<(), ProcessInputError> {
+    let mut parsed_data = parser::parse(
+        input,
+        "default_name".into(),
+        "file_name".into(),
+        "file_path".into(),
+        &[],
+        false,
+    )?
+    .unwrap();
+
+    if !parsed_data.errors.is_empty() {
+        return Err(ProcessInputError::ParseError(
+            parsed_data.errors.remove(0).error,
+        ));
+    }
+
+    language.generate_types(out, imports, parsed_data)?;
+    Ok(())
+}
 
 mod blocklisted_types {
     use std::collections::HashMap;
